Finished algorithm to sort daemon contracts with outstanding claims into contract groups. Added algorithm to update daemon contract groups in the database.

+++ Headtrip.Models/Daemon/DaemonContractGroup.cs
@@ -11,8 +11,8 @@ namespace Headtrip.Models.Daemon
         public DaemonClaim? Claim { get; set; }
         public List<DaemonContract>? Contracts { get; set; }
 
-        public int NumberOfParties { get; set; }
-        public int NumberOfPlayers { get; set; }
+        public byte NumberOfParties { get; set; }
+        public byte NumberOfPlayers { get; set; }
 
     }
 }
